fix: update and delete comment not found error handled

+++ src/services/comment.service.ts
@@ -1,6 +1,7 @@
 import type { Comment } from '@prisma/client';
 
 import type { CommentRepository } from '@/repositories/comment.repo';
+import CustomError from '@/utils/customError';
 
 export class Commentservice {
   private CommentRepository: CommentRepository;
@@ -18,10 +19,19 @@ export class Commentservice {
   }
 
   async updateComment(id: string, CommentData: any): Promise<Comment> {
+    const existingComment = await this.CommentRepository.findCommentById(id);
+    if (!existingComment) {
+      throw new CustomError('Comment not found', 404);
+    }
+
     return this.CommentRepository.updateComment(id, CommentData);
   }
 
   async deleteComment(id: string): Promise<void> {
+    const existingComment = await this.CommentRepository.findCommentById(id);
+    if (!existingComment) {
+      throw new CustomError('Comment not found', 404);
+    }
     await this.CommentRepository.deleteComment(id);
   }
 }
